src/main/java/com/solvd/testautomation/web/HomePage.java
package com.solvd.testautomation.web;

import com.solvd.testautomation.web.components.Div;
import com.zebrunner.carina.utils.config.Configuration;
import com.zebrunner.carina.webdriver.gui.AbstractPage;
import org.openqa.selenium.WebDriver;
import org.openqa.selenium.support.FindBy;

public class HomePage extends AbstractPage {

    @FindBy(xpath = "//div")
    private Div div;

    public HomePage(WebDriver driver) {
        super(driver);
    }

    @Override
    public void open() {
        openURL(Configuration.getRequired("home_url"));
    }

    public Div getHeader() {
        return div;
    }
}